requirements.txt
sparql-client==3.8